% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airport-scan.R
\name{airport_scan}
\alias{airport_scan}
\title{Scan for available wireless network (requires Wi-Fi enabled Mac)}
\usage{
airport_scan(.quiet = FALSE)
}
\arguments{
\item{.quiet}{if \code{TRUE} then no progress messages will be displayed}
}
\description{
Scan for available wireless network (requires Wi-Fi enabled Mac)
}
\note{
that we have to use the characater output since the XML output is broken on Sierra.
}
