% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-notary.R
\name{check_notarization}
\alias{check_notarization}
\title{Check application notarization info}
\usage{
check_notarization(path_to_app)
}
\arguments{
\item{path_to_app}{the path to the application or comand line binary}
}
\description{
Check application notarization info
}
