% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_sig.R
\name{check_sig}
\alias{check_sig}
\title{Check application signature/notarization information}
\usage{
check_sig(path_to_app)
}
\arguments{
\item{path_to_app}{the path to the application or comand line binary}
}
\description{
Check application signature/notarization information
}
