% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-dsstore.R
\name{find_dsstore}
\alias{find_dsstore}
\title{Find and optionally remove \code{.DS_Store} files on a locally-accessible filesystem}
\usage{
find_dsstore(path = ".", recursive = FALSE, remove = FALSE)
}
\arguments{
\item{path}{a character vector of full path names; the default corresponds to
the working directory, \code{\link[=getwd]{getwd()}}. Tilde expansion (see \code{\link[=path.expand]{path.expand()}})
is performed. Missing values will be ignored.}

\item{recursive}{logical. Should the listing recurse into directories? Default is \code{FALSE}.}

\item{remove}{if \code{TRUE} the function will attempt to \code{\link[=unlink]{unlink()}} any \code{.DS_Store}
files found. Permissions issue may cause this option to fail to delete
found \code{.DS_Store} files. Default: \code{FALSE}}
}
\description{
Provide a starting point, recursion option (similar to \code{\link[=list.files]{list.files()}}) &
whether or not to remove any \code{.DS_Store} files found and
get back a data frame of full paths to any \code{.DS_Store} files that exist
or were removed. The removal success status (if \code{remove} is \code{TRUE} will be added
as a column to the data frame).\cr\cr
If no \code{.DS_Store} files were found, an empty data frame is returned.
}
