% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-app-info.R
\name{get_app_info}
\alias{get_app_info}
\title{Retrieve iTunes info about a give app (via bundle id)}
\usage{
get_app_info(bundle_id)
}
\arguments{
\item{bundle_id}{an app bundle id}
}
\description{
Retrieve iTunes info about a give app (via bundle id)
}
