% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger}
\alias{logger}
\title{Log a message to the macOS logging system (searchable from Console.app)}
\usage{
logger(message, app = "mactheknife")
}
\arguments{
\item{message}{text to write to the log}

\item{app}{name of the "application" (makes it easier to search for your log entries)}
}
\description{
Log a message to the macOS logging system (searchable from Console.app)
}
