% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mactheknife-package.R
\docType{package}
\name{mactheknife}
\alias{mactheknife}
\alias{mactheknife-package}
\title{Read 'macOS' .DS_Store' Files}
\description{
A thin wrapper around the 'Python' 'dsstore' module
\url{https://github.com/gehaxelt/Python-dsstore} by 'Sebastian Neef'.
}
\author{
Bob Rudis (bob@rud.is)
}
