% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdls.R
\name{mdls}
\alias{mdls}
\title{Return a data frame of macOS Spotlight Metadata Attributes for the specified paths}
\usage{
mdls(paths)
}
\arguments{
\item{paths}{paths to files}
}
\description{
Return a data frame of macOS Spotlight Metadata Attributes for the specified paths
}
\examples{
mdls(
  list.files(
    pathj = file.path(find.package("mactheknife")),
    recursive = TRUE,
    full.names = TRUE
  )
)
}
