% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-plist.R
\name{read_plist}
\alias{read_plist}
\title{Read a macOS property list file}
\usage{
read_plist(path)
}
\arguments{
\item{path}{path to \code{plist} file}
}
\value{
\code{list} with property list contents or \code{NULL} if there was an error
reading the property list.
}
\description{
Read a macOS property list file
}
