% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve-alias.R
\name{resolve_alias}
\alias{resolve_alias}
\title{Resolve macOS Alias Files to POSIX path strings}
\usage{
resolve_alias(path)
}
\arguments{
\item{path}{path to the target alias}
}
\value{
character vector
}
\description{
Resolve macOS Alias Files to POSIX path strings
}
