% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-knowledgeC.R
\name{src_knowledgeC}
\alias{src_knowledgeC}
\title{Source the `knowledgeC.db` on the local system}
\usage{
src_knowledgeC(
  knowledgeC_location = "~/Library/Application Support/Knowledge/knowledgeC.db"
)
}
\arguments{
\item{knowledgeC_location}{location of the `knowledgeC.db` file (defaults to "`~/Library/Application Support/Knowledge/knowledgeC.db`")}
}
\description{
Source the `knowledgeC.db` on the local system
}
