% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mactheknife-package.R
\docType{package}
\name{mactheknife}
\alias{mactheknife}
\alias{mactheknife-package}
\title{Read 'macOS' .DS_Store' Files}
\description{
A set of tools/methods and data that are geared towards the 'macOS' ecosystem.
}
\author{
Bob Rudis (bob@rud.is)
}
