% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-dsstore.R
\name{read_dsstore}
\alias{read_dsstore}
\title{Read a \code{.DS_Store} file}
\usage{
read_dsstore(path)
}
\arguments{
\item{path}{a path to a valid \code{.DS_Store} file (\code{\link[=path.expand]{path.expand()}} will be called)}
}
\value{
a character vector of filenames in the \code{.DS_Store} file or
a length 0 character vector if no parseable data was found
}
\description{
Read a \code{.DS_Store} file
}
\examples{
read_dsstore(system.file("extdat", "DS_Store.ctf", package = "mactheknife"))
}
